backend/requirements.txt
django==3.2.7
django-grappelli
django-hashedfilenamestorage
django-filter
djangorestframework
drf-yasg
drf-yasg[validation]
gunicorn
django-heroku
Pillow
django-nested-admin
django-admin-sortable2